refactor: add problem input middleware (#73)
Use input case table interface.
