+++ routers/day/index.js
@@ -1,9 +1,10 @@
 const express = require('express');
-const router = express.Router();
-const fs = require('fs/promises');
 const validateChallengeNumber = require('#middlewares/validate-challenge-number');
 const validateAccessTime = require('#middlewares/validate-access-time');
 const dayPageRenderer = require('#middlewares/day-page-renderer');
+const problemInputResponse = require('#middlewares/problem-input-response');
+
+const router = express.Router();
 
 router.get('/:day',
   validateChallengeNumber,
@@ -13,19 +14,6 @@ router.get('/:day',
 router.get('/:day/input',
   validateChallengeNumber,
   validateAccessTime,
-  async (req, res, next) => {
-    if (!res.locals.signedIn) {
-      throw new PageNotFoundError();
-    }
-
-    const day = req.params.day;
-    const problemSetNumber = day + '0';
-    const inputCasesIdx = parseInt(day) - 1;
-    const inputCase = req.session.user.inputCases[inputCasesIdx];
-    const buffer = await fs.readFile('./challenges/inputs/' + problemSetNumber + '/' + inputCase + '.txt');
-    const input = buffer.toString();
-    res.type('text/plain');
-    res.send(input);
-  });
+  problemInputResponse);
 
 module.exports = router;
